fix data type for product price

+++ database/migrations/2022_06_23_015216_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('price');
+            $table->decimal('price');
             $table->integer('quantity');
             $table->boolean('is_active')->default(1);
             $table->timestamps();
